Update button styles for mouse over and checked states
Modified `ControlTemplate.Triggers` in `ButtonStyle.xaml` to change the `Foreground` color of `Icon` and `Lbl` from `White` to `Gray` on mouse over. Removed the background color setting for `B` during mouse over. The `IsChecked` trigger remains unchanged, maintaining the `Background` of `B` as `Silver` and `Icon` as `White` when checked.